test/backend/requirements-test.txt
# Dependencias para testing del backend Django

# Testing framework principal
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0

# Testing utilities
factory-boy==3.3.0
faker==20.1.0
model-bakery==1.17.0

# HTTP testing
requests-mock==1.11.0
responses==0.24.1

# Database testing
pytest-postgresql==5.0.0
pytest-factoryboy==2.6.0

# Coverage y reporting
coverage==7.3.2
pytest-html==4.1.1
pytest-json-report==1.5.0

# Linting y formatting para tests
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Debugging
pytest-pdb==0.2.0
pytest-sugar==0.9.7

# Performance testing
pytest-benchmark==4.0.0

# Parallel testing
pytest-parallel==0.1.1

# Django specific testing
django-test-plus==2.2.2
django-webtest==1.9.11

# API testing
django-rest-framework==3.14.0
djangrestframework-simplejwt==5.3.0

# Environment management
python-decouple==3.8

# Time/Date mocking
freezegun==1.2.2

# Memory profiling
memory-profiler==0.61.0

# Load testing (opcional)
locust==2.17.0